create archive method and zip tasks
